Forward port dbus-java recipes and some more openjdk-7 fixes

From meta-java rev: 716bae8a8b070b88fa91c30e03fc09e70310091a

+++ meta-java/recipes-core/junit/junit_3.8.2.bb
@@ -14,8 +14,9 @@ do_unpackpost() {
 	mkdir -p src
 
 	# Prevent deletion by do_removebinaries.
-	mv src.jar src.zip
-
+	if [ -e src.jar ]; then
+		mv src.jar src.zip
+	fi
 	unzip src.zip -d src
 }
